Move PrintRegisters&PrintRegister to helpers package
